Disable `type` key at the top level of templates
Since the `type` are only for organization, and various forks of GAMER
are owned by individuals, `type` are commented out in the templates.
```
Issue types, sub-issues, and advanced issue search are currently in
public preview for organizations. See the GitHub Blog.
https://github.blog/changelog/2025-01-13-evolving-github-issues-public-preview/
```
And after our discussion in the lab, we decided not to add `project` and
`type` because it might create more problems in the forked repositories.