Fix for .cu to support negative passive scalars
Fixed by passing the global variable `PassiveVar_Floor` as a parameter
to the kernel functions, layer by layer.
This include editing:
- 6  function bodies
- 42 headers & calls
- 8  comments
And `Hydro_IsUnphysical` has now been fixed yet. Because the number of
its references is too large, and I need a seperate commit to fix it.
Actually, this leads to PR #443. And I rebase this feature branch after
the PR is merged.